graph printing deduplicated and made accessible outside CLI
